+++ src/common/src/utils.ts
@@ -368,10 +368,10 @@ export class Utils {
       }
     })
   }
-  proceedWithAutoInstall(params: IAutoParam, setup: SetupEnvironment) {
+  proceedWithAutoInstall(params: IAutoParam, setup: SetupEnvironment, purge = true) {
     return new Observable((observer) => {
       // console.log('hzn_css', pEnv.HZN_CSS, typeof pEnv.HZN_CSS, Boolean(pEnv.HZN_CSS))
-      this.purgeManagementHub() // Leverage this function to cleanup and install prerequisites, maynot need preInstallHznCli anymore
+      this.purgeManagementHub(purge) // Leverage this function to cleanup and install prerequisites, maynot need preInstallHznCli anymore
       .subscribe({
         complete: () => {
           const pEnv: any = process.env;
@@ -428,6 +428,7 @@ export class Utils {
         observer.next('Please provide --config_file name')
         observer.complete()
       } else if(setup == SetupEnvironment.autoSetupAllInOne || setup == SetupEnvironment.autoSetupCliInContainer || setup == SetupEnvironment.autoSetupAnaxInContainer || setup == SetupEnvironment.autoSetupContainer || setup == SetupEnvironment.autoSetupOpenHorizonMesh) {
+        const purge = setup != SetupEnvironment.autoSetupOpenHorizonMesh;
         let configJson
         this.updateConfig(configFile)
         .subscribe({
@@ -441,7 +442,7 @@ export class Utils {
             Object.keys(org).forEach((key) => {
               pEnv[key] = org[key]
             })
-            this.proceedWithAutoInstall(params, setup)
+            this.proceedWithAutoInstall(params, setup, purge)
             .subscribe({
               complete: () => {
                 observer.next('')
@@ -957,8 +958,8 @@ export class Utils {
     const arg = `docker exec horizon1 rm -rf /var/horizon/anax.db`
     return this.shell(arg)
   }
-  purgeManagementHub() {
-    if(os.arch() == 'x64' || process.platform == 'darwin') {
+  purgeManagementHub(purge: boolean) {
+    if(purge && (os.arch() == 'x64' || process.platform == 'darwin')) {
       const arg = `curl -sSL https://raw.githubusercontent.com/open-horizon/devops/master/mgmt-hub/deploy-mgmt-hub.sh --output deploy-mgmt-hub.sh && chmod +x deploy-mgmt-hub.sh && sudo ./deploy-mgmt-hub.sh -PS && sudo rm -rf /tmp/horizon-all-in-1`
       return this.shell(arg)  
     } else {
@@ -980,7 +981,7 @@ export class Utils {
   }
   unregisterMeshAgent() {
     const pEnv = process.env;
-    const arg = `curl -sL --insecure -u $HZN_ORG_ID/$HZN_EXCHANGE_USER_AUTH -X DELETE ${pEnv.MESH_ENDPOINT}/v1/orgs/${pEnv.HZN_ORG_ID}/nodes/${pEnv.HZN_DEVICE_ID}`;
+    const arg = `curl -sL --insecure -u ${pEnv.HZN_ORG_ID}/${pEnv.HZN_EXCHANGE_USER_AUTH} -X DELETE ${pEnv.MESH_ENDPOINT}/v1/orgs/${pEnv.HZN_ORG_ID}/nodes/${pEnv.HZN_DEVICE_ID}`;
     return this.shell(arg);
   }
   setupOpenHorizonMesh(params: IAutoParam, anax: string) {
